Separate UIs for analysis vs input
