+++ .streamlit/config.toml
@@ -1,3 +1,6 @@
+[client]
+showSidebarNavigation = false
+
 [server]
 # Allow large .h5ad datasets (10GB). Default is 200MB.
 maxUploadSize = 10240
